Module libray::math

Math module

Contains fundamental math type and methods.

Structs

Intersection

Aggregation of time and object that was intersected

Matrix

Matrix representation

Ray

Represents an individual ray

Sphere

Represents a sphere object

Tuple

3 Dimensional Tuple struct representing points or vectors.

Functions

normal_at

Returns the normal vector of given sphere at given point

reflect

Returns the reflected vector of given vector around a normal

rotation_x

Returns a rotation matrix around the x axis

rotation_y

Returns a rotation matrix around the y axis

rotation_z

Returns a rotation matrix around the z axis

scaling

Returns a scling matrix with the given scale units

skewing

Returns a skewing matrix

translation

Returns a translation matrix with the given translation units